engine/pyproject.toml
[project]
name = "nexus-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "peft>=0.18.1",
    "protobuf>=6.33.6",
    "pydantic>=2.12.5",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
